backend/requirements.txt
Django
django-filter
django-cors-headers
djangorestframework
psycopg2-binary
django-debug-toolbar
gunicorn
flower
sentry-sdk
requests
arrow
drf-spectacular
python-dotenv
dj-rest-auth
pillow
